Move contrib.indexing to ops.indexing
