+++ examples/hmm_enum.py
@@ -58,11 +58,11 @@ import jax.numpy as jnp
 
 import numpyro
 from numpyro.contrib.control_flow import scan
-from numpyro.contrib.indexing import Vindex
 import numpyro.distributions as dist
 from numpyro.examples.datasets import JSB_CHORALES, load_dataset
 from numpyro.handlers import mask
 from numpyro.infer import HMC, MCMC, NUTS
+from numpyro.ops.indexing import Vindex
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
