+++ test/contrib/test_funsor.py
@@ -17,9 +17,9 @@ from numpyro.contrib.control_flow import scan
 from numpyro.contrib.funsor import config_enumerate, enum, markov, to_data, to_funsor
 from numpyro.contrib.funsor.enum_messenger import NamedMessenger, plate as enum_plate
 from numpyro.contrib.funsor.infer_util import log_density
-from numpyro.contrib.indexing import Vindex
 import numpyro.distributions as dist
 from numpyro.infer import MCMC, NUTS, init_to_median
+from numpyro.ops.indexing import Vindex
 from numpyro.primitives import _PYRO_STACK
